+++ lib/services/remote_directory_picker_window_service.dart
@@ -1,0 +1,2 @@
+export 'remote_directory_picker_window_service_io.dart'
+    if (dart.library.html) 'remote_directory_picker_window_service_web.dart';